requirements.txt
aiofiles==0.7.0; python_version >= "3.6" and python_version < "4.0"
async-exit-stack==1.0.1; python_version >= "3.5"
async-generator==1.10; python_version >= "3.5"
attrs==20.3.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7"
certifi==2021.5.30; python_version >= "3.6"
click==7.1.2; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
dnspython==2.1.0; python_version >= "3.6"
email-validator==1.1.2
faker==8.6.0; python_version >= "3.6"
fastapi==0.65.1; python_version >= "3.6"
greenlet==1.1.0; python_version >= "3" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3"
h11==0.12.0; python_version >= "3.6"
httpcore==0.13.3; python_version >= "3.6"
httpx==0.18.1; python_version >= "3.6"
idna==2.10; python_version >= "3.6" and python_full_version < "3.0.0" or python_version >= "3.6" and python_full_version >= "3.4.0"
jinja2==2.11.3; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
jsf==0.3.2; python_version >= "3.7"
jsonref==0.2
jsonschema==3.2.0; python_version >= "3.7"
markupsafe==2.0.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
pydantic==1.8.2; python_full_version >= "3.6.1" and python_version >= "3.7"
pyrsistent==0.17.3; python_version >= "3.7"
python-dateutil==2.8.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.7"
python-dotenv==0.17.1
python-multipart==0.0.5
rfc3986==1.5.0; python_version >= "3.6"
rstr==2.2.6; python_version >= "3.7"
six==1.16.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.7"
smart-open==5.1.0; python_version >= "3.7"
sniffio==1.2.0; python_version >= "3.6"
sqlalchemy==1.4.17; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
starlette==0.14.2; python_version >= "3.6"
text-unidecode==1.3; python_version >= "3.7"
typer==0.3.2; python_version >= "3.6"
typing-extensions==3.10.0.0; python_full_version >= "3.6.1" and python_version >= "3.6"
uvicorn==0.13.4
